refacto: split each section into a specific file. create abstract to centralize useful functions.
