+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/dashboard/CustomizableCell.kt
@@ -3,7 +3,7 @@ package fr.gouv.cacem.monitorenv.domain.entities.dashboard
 import org.apache.poi.xwpf.usermodel.XWPFDocument
 import org.apache.poi.xwpf.usermodel.XWPFTableCell
 
-interface CustomizableCell {
+fun interface CustomizableCell {
     fun customizeValueCell(
         rowIndex: Int,
         cell: XWPFTableCell,
