+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/dashboard/DetailRenderable.kt
@@ -1,5 +1,5 @@
 package fr.gouv.cacem.monitorenv.domain.entities.dashboard
 
-interface DetailRenderable {
+fun interface DetailRenderable {
     fun buildDetailsRows(): List<List<String>>
 }
