+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/entities/dashboard/EditableBriefRecentPerUnitActivityEntity.kt
@@ -1,7 +1,6 @@
 package fr.gouv.cacem.monitorenv.domain.entities.dashboard
 
 data class EditableBriefRecentPerUnitActivityEntity(
-    val controlUnitId: Int,
-    val recentControls: List<EditableBriefRecentControlActivityEntity>,
+    val id: Int,
     val image: String? = null,
 )
