+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/infrastructure/file/dashboard/IBriefFileWriter.kt
@@ -1,0 +1,11 @@
+package fr.gouv.cacem.monitorenv.infrastructure.file.dashboard
+
+import fr.gouv.cacem.monitorenv.domain.entities.dashboard.BriefEntity
+import org.apache.poi.xwpf.usermodel.XWPFDocument
+
+fun interface IBriefFileWriter {
+    fun createSection(
+        brief: BriefEntity,
+        document: XWPFDocument,
+    )
+}
