+++ backend/src/test/kotlin/fr/gouv/cacem/monitorenv/infrastructure/api/endpoints/bff/v1/DashboardsITests.kt
@@ -379,7 +379,7 @@ class DashboardsITests {
                 vigilanceAreas = emptyList(),
                 regulatoryAreas = emptyList(),
                 nearbyUnits = emptyList(),
-                recentActivity = EditableBriefRecentActivityEntity(null, null, "toto", emptyList()),
+                recentActivity = EditableBriefRecentActivityEntity(null, null, "toto", emptyList(), emptyList()),
             )
         given(createBrief.execute(brief)).willReturn(
             BriefFileEntity(
